+++ examples/v2/microsoft-teams-integration/GetApiHandle.rb
@@ -1,0 +1,5 @@
+# Get handle information returns "OK" response
+
+require "datadog_api_client"
+api_instance = DatadogAPIClient::V2::MicrosoftTeamsIntegrationAPI.new
+p api_instance.get_api_handle("handle_id")