+++ lib/datadog_api_client/v2/models/fastly_services_response.rb
@@ -38,7 +38,7 @@ module DatadogAPIClient::V2
     # @!visibility private
     def self.openapi_types
       {
-        :'data' => :'Array<FastlyServiceResponse>'
+        :'data' => :'Array<FastlyServiceData>'
       }
     end