+++ lib/datadog_api_client/v2/models/security_monitoring_standard_rule_create_payload.rb
@@ -45,6 +45,9 @@ module DatadogAPIClient::V2
     # Queries for selecting logs which are part of the rule.
     attr_reader :queries
 
+    # Reference tables for the rule.
+    attr_accessor :reference_tables
+
     # Tags for generated signals.
     attr_accessor :tags
 
@@ -68,6 +71,7 @@ module DatadogAPIClient::V2
         :'name' => :'name',
         :'options' => :'options',
         :'queries' => :'queries',
+        :'reference_tables' => :'referenceTables',
         :'tags' => :'tags',
         :'third_party_cases' => :'thirdPartyCases',
         :'type' => :'type'
@@ -86,6 +90,7 @@ module DatadogAPIClient::V2
         :'name' => :'String',
         :'options' => :'SecurityMonitoringRuleOptions',
         :'queries' => :'Array<SecurityMonitoringStandardRuleQuery>',
+        :'reference_tables' => :'Array<SecurityMonitoringReferenceTable>',
         :'tags' => :'Array<String>',
         :'third_party_cases' => :'Array<SecurityMonitoringThirdPartyRuleCaseCreate>',
         :'type' => :'SecurityMonitoringRuleTypeCreate'
@@ -145,6 +150,12 @@ module DatadogAPIClient::V2
       if attributes.key?(:'queries')
         if (value = attributes[:'queries']).is_a?(Array)
           self.queries = value
+        end
+      end
+
+      if attributes.key?(:'reference_tables')
+        if (value = attributes[:'reference_tables']).is_a?(Array)
+          self.reference_tables = value
         end
       end
 
@@ -272,6 +283,7 @@ module DatadogAPIClient::V2
           name == o.name &&
           options == o.options &&
           queries == o.queries &&
+          reference_tables == o.reference_tables &&
           tags == o.tags &&
           third_party_cases == o.third_party_cases &&
           type == o.type
@@ -282,7 +294,7 @@ module DatadogAPIClient::V2
     # @return [Integer] Hash code
     # @!visibility private
     def hash
-      [cases, filters, has_extended_title, is_enabled, message, name, options, queries, tags, third_party_cases, type].hash
+      [cases, filters, has_extended_title, is_enabled, message, name, options, queries, reference_tables, tags, third_party_cases, type].hash
     end
   end
 end